+++ rs/http_endpoints/public/src/lib.rs
@@ -31,6 +31,7 @@ pub use call::{
 pub use common::cors_layer;
 pub use query::QueryServiceBuilder;
 pub use read_state::canister::{CanisterReadStateService, CanisterReadStateServiceBuilder};
+pub use read_state::subnet::SubnetReadStateServiceBuilder;
 
 use crate::{
     catch_up_package::CatchUpPackageService,
@@ -46,7 +47,6 @@ use crate::{
         STATUS_SUCCESS,
     },
     pprof::{PprofFlamegraphService, PprofHomeService, PprofProfileService},
-    read_state::subnet::SubnetReadStateService,
     status::StatusService,
     tracing_flamegraph::TracingFlamegraphService,
 };
@@ -380,11 +380,10 @@ pub fn start_server(
     .with_malicious_flags(malicious_flags)
     .build_router();
 
-    let subnet_read_state_router = SubnetReadStateService::new_router(
-        Arc::clone(&health_status),
-        Arc::clone(&delegation_from_nns),
-        state_reader.clone(),
-    );
+    let subnet_read_state_router =
+        SubnetReadStateServiceBuilder::builder(delegation_from_nns.clone(), state_reader.clone())
+            .with_health_status(health_status.clone())
+            .build_router();
     let status_router = StatusService::build_router(
         log.clone(),
         nns_subnet_id,
@@ -1091,6 +1090,7 @@ async fn get_random_node_from_nns_subnet(
 
 #[cfg(test)]
 mod tests {
+    use crate::read_state::subnet::SubnetReadStateService;
     use bytes::Bytes;
     use futures_util::{future::select_all, stream::pending, FutureExt};
     use http::{